+++ HISTORY.md
@@ -1,5 +1,9 @@
 History of features added to LiChess Tools in time (reversed order)
 
+18 Nov
+  - Interactive lesson Preview mode sticks between page reloads as well
+  - Option to open links from move list comments that lead to studies in the same window
+
 16 Nov
   - Analysis Wiki FEN search
   - Copy PGN tohere parameter